requirements.txt
flask==3.1.3
flask-sqlalchemy==3.1.1
bcrypt==5.0.0
pillow==12.2.0
pyjwt==2.12.1
werkzeug==3.1.8
gunicorn==21.2.0
psycopg2-binary==2.9.10
